+++ HISTORY.rst
@@ -5,14 +5,20 @@
 History
 =======
 
-current - 2018-05-18 - 0.00Mb
+current - 2018-10-27 - 0.00Mb
 =============================
 
-* `7`: implement read_json (2018-05-17)
+* `13`: implement concat for axis=0,1 (2018-10-26)
+* `12`: add groupby_streaming (2018-10-26)
+* `11`: add method add_column (2018-10-26)
+* `9`: head is very slow (2018-10-26)
+* `10`: plan B to bypass a bug in pandas about read_csv when iterator=True --> closed, pandas has a weird behaviour when names is too small compare to the number of columns (2018-10-26)
+* `8`: fix pandas_streaming for pandas 0.23.1 (2018-07-31)
 
 0.1.87 - 2018-05-17 - 0.02Mb
 ============================
 
+* `7`: implement read_json (2018-05-17)
 * `6`: add pandas_groupby_nan from pyensae (2018-05-17)
 
 0.1.66 - 2018-02-05 - 0.02Mb